Create README.md
Small arcade type videogame, which will be adapted to a story mode. The operator will have the opportunity to select a character of his choice, during the course of the story, the player must face various challenges, in turn the character can improve some of its attributes by performing specific actions. This game employs turn-based action, so the player must perform an action, and then wait for the other character to act, and then act again. The story of the game has not been written, so it is still unknown what the characters will be like, and how they will interact when confronted.  The game is intended to run on the terminal, implementing a small variety of sprites that will act as the characters.